fix: 修复 CI 编译错误
- RemoteInferenceClient: b/cv 局部 val 避免 smart cast 不允许
- RemoteInferenceServer: 补 import AtomicInteger/AtomicReference
- RemoteInferenceServer: 改 ?: run{continue} 为 if null continue
- RemoteInferenceServer: list 类型标注为 MutableList
- FloatService: elapsed 显式 toInt() 避免 Long/Int LUB 推断

+++ app/src/main/java/com/yolotouchhelp/aimbot/service/FloatService.kt
@@ -843,7 +843,7 @@ class FloatService : Service() {
                         framesInWindow += (cur - lastFrameId).coerceAtLeast(0)
                         lastFrameId = cur
                     }
-                    val elapsed = now - lastFpsT
+                    val elapsed = (now - lastFpsT).toInt()
                     val fps: Int = if (elapsed >= 1000) {
                         val f = framesInWindow * 1000 / elapsed
                         framesInWindow = 0
